Changes to column specification syntax.
The path-query syntax is expanded here to include a view
specification as a sequence of <view> elements. This allows
for functions to be applied to the paths without changing the path
syntax itself (eg: <view fn="COUNT">Some.path.here</view>). The
PathQuery can then manage a map from path -> fn.